Move metaprogramming to new class
